Makefile.in: Update the libtool flags used to build the pyunbound extension module
This patch changes the flags passed to libtool when building the
pyunbound Python extension module. Building a Python extension module is
substantially different from building a shared library with libtool.

Specifically,

1) Use "-avoid-version" instead of "-version-info". The Python module
loader does not look for the version-suffixed filenames, so avoid
generating them. (E.g., Python looks for "_unbound.so", not
"_unbound.so.2" or "_unbound.so.2.3.3".)

2) Use "-shared", which produces only the dynamic library. The Python
module loader is dlopen() based and cannot utilize ".a" files.

3) Don't add "$(LIBS)" to the link flags. This pulls in extra libraries
that are needed for other libraries and programs built by the Makefile,
but aren't needed for the Python extension module. E.g., on my system,
$(LIBS) pulls in libdl, libpython2.7, and libcrypto. AFAICT, the Python
extension module _unbound.so should only be linked against libc and
libunbound. (From looking at the symbols referenced by _unbound.so, no
symbols from libdl and libcrypto are used, and the Py* symbols are
already loaded by the Python runtime, leaving only C standard library
and libunbound symbols.)

This still results in an overlinked library; libtool still unfortunately
links _unbound.so to all the libraries that libunbound depends on.