Don't log successful setrlimit() call to console
This message is printed on a successful rlimits change in the number of
open file descriptors. Since it occurs before unbound has daemonized,
the message goes directly to the console and interferes with the init
script output, e.g.:

    root@chase:~# /etc/init.d/unbound start
    Starting recursive DNS server: unbound[1415750323] unbound[8070:0] info: increased limit(open files) from 1024 to 4152
    .
    root@chase:~#

Since this is a relatively innocuous message, just delete it.
setrlimit() failures will still be logged.