debian/unbound.init: Rewrite to use package-helper script
Rewrite the sysvinit script, relying on factored out functionality in
the package-helper script to setup the chroot, to add ourself to
resolvconf, and to update the root trust anchor file.

Remove the handling of the UNBOUND_ENABLE variable entirely. This is not
the right way to handle enabling or disabling an installed daemon.