debian/control: Build-Depend on python '-dev' packages, not '-all-dev' (Closes: #864334, #866770)
In Ubuntu, we've begun the transition from python3.5 to python3.6, which in
the first stages involves enabling python3.6 as a non-default, supported
python3 version.

The unbound package build-depends on python3-all-dev, so was in our list of
packages needing rebuilt; however, it only references py3versions -d in
debian/rules, so does not build support for any non-default python versions.

Please find attached the small patch to debian/control, applied in Ubuntu,
to switch from -all-dev to -dev, which optimizes the build and makes it
clear how this package relates to python transitions.

[edmonds@debian.org: Imported patch description from BTS.]